Refactor : move non-yearly fields to partnership
