+++ templates/partnerships/includes/partnership_form_course.html
@@ -45,7 +45,7 @@
     {% bootstrap_field form_year.education_fields %}
   </div>
   <div class="col-xs-12">
-    {% bootstrap_field form_year.subtype field_class="form-inline" %}
+    {% bootstrap_field form.subtype field_class="form-inline" %}
   </div>
   <div class="col-xs-12">
     {% bootstrap_field form_year.education_levels %}
@@ -57,7 +57,7 @@
     {% bootstrap_field form_year.offers %}
   </div>
   <div class="col-xs-12">
-    {% bootstrap_field form_year.description %}
+    {% bootstrap_field form.description %}
   </div>
   <div class="col-xs-12">
     {% bootstrap_field form.tags %}
