+++ templates/partnerships/includes/partnership_form_general.html
@@ -44,16 +44,16 @@
     {% bootstrap_field form_year.education_fields %}
   </div>
   <div class="col-xs-12">
-    {% bootstrap_field form_year.subtype field_class="form-inline" %}
+    {% bootstrap_field form.subtype field_class="form-inline" %}
   </div>
   <div class="col-xs-12">
     {% bootstrap_field form_year.education_levels %}
   </div>
   <div class="col-xs-12">
-    {% bootstrap_field form_year.missions field_class="form-inline" %}
+    {% bootstrap_field form.missions field_class="form-inline" %}
   </div>
   <div class="col-xs-12">
-    {% bootstrap_field form_year.description %}
+    {% bootstrap_field form.description %}
   </div>
   <div class="col-xs-12">
     {% bootstrap_field form.tags %}
