+++ templates/partnerships/includes/partnership_form_project.html
@@ -48,22 +48,22 @@
     {% bootstrap_field form_year.education_levels %}
   </div>
   <div class="col-xs-12">
-    {% bootstrap_field form_year.missions field_class="form-inline" %}
+    {% bootstrap_field form.missions field_class="form-inline" %}
   </div>
   <div class="col-xs-12">
     {% bootstrap_field form_year.funding %}
   </div>
   <div class="col-xs-12">
-    {% bootstrap_field form_year.project_title %}
+    {% bootstrap_field form.project_title %}
   </div>
   <div class="col-xs-12">
-    {% bootstrap_field form_year.id_number %}
+    {% bootstrap_field form.id_number %}
   </div>
   <div class="col-xs-12">
-    {% bootstrap_field form_year.ucl_status %}
+    {% bootstrap_field form.ucl_status %}
   </div>
   <div class="col-xs-12">
-    {% bootstrap_field form_year.description %}
+    {% bootstrap_field form.description %}
   </div>
   <div class="col-xs-12">
     {% bootstrap_field form.tags %}
