+++ tests/factories/partnership.py
@@ -50,3 +50,8 @@ class PartnershipFactory(factory.DjangoModelFactory):
                 ])
             elif extracted:
                 obj.years.set(extracted)
+
+    @factory.post_generation
+    def missions(obj, create, extracted, **kwargs):
+        if create and extracted is not None:
+            obj.missions.set(extracted)
