+++ tests/views/partnership/test_partnership_course.py
@@ -86,8 +86,8 @@ class PartnershipCreateCourseViewTest(TestCase):
             'year-start_academic_year': cls.start_academic_year.pk,
             'year-end_academic_year': cls.end_academic_year.pk,
             'year-funding_type': FundingTypeFactory().pk,
-            'year-missions': PartnershipMissionFactory().pk,
-            'year-subtype': PartnershipSubtypeFactory().pk,
+            'missions': PartnershipMissionFactory().pk,
+            'subtype': PartnershipSubtypeFactory().pk,
         }
 
     def test_get_view_as_adri(self):
@@ -173,11 +173,11 @@ class PartnershipUpdateCourseViewTest(TestCase):
             'year-entities': [],
             'year-offers': [],
             'year-funding_type': FundingTypeFactory().pk,
-            'year-missions': [
+            'missions': [
                 PartnershipMissionFactory().pk,
                 PartnershipMissionFactory().pk,
             ],
-            'year-subtype': PartnershipSubtypeFactory().pk,
+            'subtype': PartnershipSubtypeFactory().pk,
         }
 
     def test_get_own_partnership_as_adri(self):
